refactor(nim-serving): move k8s-calling functions from accounts/utils to accounts/api
Follow-up from PR #7466 review feedback: functions that make k8s API
calls belong in api.ts files, not utils.ts. This applies the same
pattern already established in the images/ directory to the accounts/
directory, keeping utils.ts for pure utility functions only.
